Add a HandleCache
