+++ src/vertex_layout_factory.h
@@ -6,10 +6,6 @@
 
 #include "resource_array.h"
 
-/*
-bool CreateVertexLayout(const std::vector<D3D11_INPUT_ELEMENT_DESC>& input_layout, ID3DBlob* shader_blob, ID3D11Device* device, ID3D11InputLayout** vertex_layout);
-*/
-
 using VertexLayoutHandleType = ResourceArray::HandleType;
 
 ResourceArray::HandleType CreateVertexLayout(const std::vector<D3D11_INPUT_ELEMENT_DESC>& input_layout, ID3DBlob* shader_blob, ID3D11Device* device);
